backend/requirements.txt
accelerate==1.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
attrs==25.1.0
av==14.2.0
babel==2.17.0
beautifulsoup4==4.13.3
binaryornot==0.4.4
bitsandbytes==0.45.3
bleach==6.2.0
blessed==1.20.0
blinker==1.9.0
blis==1.2.0
Brotli==1.1.0
browserforge==1.2.3
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
cookiecutter==2.6.0
Crawl4AI==0.6.3
crawlee==0.6.3
crawler==0.0.2
cryptography==44.0.2
cssselect==1.2.0
csvw==3.5.1
ctranslate2==4.5.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
curl_cffi==0.9.0
cymem==2.0.11
DateTime==5.5
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
diffusers==0.32.2
distro==1.9.0
dlinfo==2.0.0
dnspython==2.7.0
docopt==0.6.2
docutils==0.21.2
editor==1.6.6
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
enum34==1.1.10
espeakng-loader==0.2.4
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.2.0
fake-http-header==0.3.5
fake-useragent==2.2.0
faster-whisper==1.1.1
fastjsonschema==2.21.1
feedparser==6.0.11
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.17.0
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.2.0
future==1.0.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
globals==0.3.36
googlenewsdecoder==0.1.7
greenlet==3.1.1
grpcio==1.67.1
grpcio-tools==1.67.1
h11==0.14.0
h2==4.2.0
h3==4.2.1
halo==0.0.31
hpack==4.1.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.2
humanfriendly==10.0
humanize==4.12.3
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
inquirer==3.4.0
ipykernel==6.29.5
ipython==8.33.0
ipywidgets==8.1.5
isodate==0.7.2
isoduration==20.11.0
itsdangerous==2.2.0
jaro-winkler==2.0.3
jedi==0.19.2
Jinja2==3.1.6
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keybert==0.9.0
kokoro==0.8.4
langcodes==3.5.0
langdetect==1.0.9
language-tags==1.2.0
language_data==1.3.0
litellm==1.69.3
log-symbols==0.0.14
loguru==0.7.3
lxml==5.3.1
lxml_html_clean==0.4.2
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
milvus-lite==2.4.11
minio==7.2.15
misaki==0.8.4
mistune==3.1.2
more-itertools==10.6.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.12
nanoid==2.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
num2words==0.5.14
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnxruntime==1.20.1
openai==1.75.0
openwakeword==0.6.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parsel==1.10.0
parso==0.8.4
pexpect==4.9.0
pgvector==0.3.6
phonemizer-fork==3.3.2
pika==1.3.2
pillow==10.4.0
platformdirs==4.3.6
playwright==1.50.0
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.4
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pvporcupine==1.9.5
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic-settings==2.6.1
pydantic_core==2.27.2
pydub==0.25.1
pyee==12.1.1
PyGetWindow==0.0.9
Pygments==2.19.1
PyJWT==2.10.1
pymilvus==2.5.5
pymongo==4.11.2
PyMsgBox==1.0.9
pyOpenSSL==25.0.0
pyparsing==3.2.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
python-slugify==8.0.4
python3-xlib==0.15
python_helper==0.3.74
pytweening==1.2.0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
rank-bm25==0.2.2
ratelim==0.1.6
rdflib==7.1.3
readability-lxml==0.8.4.1
readchar==4.2.1
RealtimeSTT==0.3.95
redis==6.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
runs==1.2.2
safetensors==0.5.3
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.1
segments==2.3.0
selectolax==0.3.29
Send2Trash==1.8.3
sentence-transformers==3.4.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcollections==2.1.0
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.6
spacy==3.8.4
spacy-curated-transformers==0.3.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spinners==0.0.24
srsly==2.5.1
stack-data==0.6.3
sympy==1.13.1
termcolor==2.5.0
terminado==0.18.1
text-unidecode==1.3
tf-playwright-stealth==1.1.2
tflite-runtime==2.14.0
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.9.0
timezonefinder==6.5.8
tinycss2==1.4.0
tldextract==5.1.3
tokenizers==0.21.0
tomli==2.2.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
triton==3.2.0
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.3.0
w3lib==2.3.1
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
webrtcvad-wheels==2.0.14
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.2
xmod==1.8.1
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zope.interface==7.2
zstandard==0.23.0
